changed structure of index and changed when to push to json file

+++ indexer.py
@@ -1,5 +1,6 @@
 import json
 import nltk
+from main import globalIndex
 
 def createIndex(tokens: [str], html: int):
     """creates the index of all tokens with list of postings
@@ -10,22 +11,25 @@ def createIndex(tokens: [str], html: int):
     # create a frequency dictionary for all the tokens
     freqDict = nltk.FreqDist(tokens)
 
-    # first open indexer file and read its content
-    with open('index.json', 'w+') as jsonFile:
-        index = json.loads(jsonFile)
-        # go through all tokens and add to the indexer
-        for token in freqDict:
-            postings = [{html: [0]}, 0]
-            if token in index:
-                # list of postings
-                postings = index[token]
-                # add key into document dictionary
-                frequency = postings[0][html]
-                updatedFrequency = frequency + 1
-                postings[0][html][0] = updatedFrequency # this will have an updated list
-            index[token] = postings
-            # start from beginning of file
+    # go through all tokens and add to the global indexer
+    for token in freqDict:
+        postings = [{html: [freqDict[token]]}, 0]
+        if token in globalIndex:
+            # list of postings
+            postings = globalIndex[token]
+            # add key into document dictionary
+            frequency = postings[0][html][0]
+            updatedFrequency = frequency + freqDict[token]
+            postings[0][html][0] = updatedFrequency # this will have an updated list
+        globalIndex[token] = postings
+
+    # if index is a multiple of 100, we dump into jsonFile and clear the global index
+    if html % 100 == 0:
+        with open('index.json', 'r+') as jsonFile:
             jsonFile.seek(0)
             # add to jsonFile and
-            json.dump(index, jsonFile)
+            json.dump(globalIndex, jsonFile)
             jsonFile.truncate()
+        globalIndex.clear()
+
+
